requirements.txt
attrs==25.3.0
blinker==1.9.0
click==8.2.1
dnspython==2.7.0
Flask==3.1.1
flask-cors==6.0.1
Flask-PyMongo==3.0.1
gunicorn==23.0.0
iniconfig==2.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
more-itertools==10.7.0
packaging==25.0
pluggy==1.6.0
py==1.11.0
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.13.2
pytest==8.4.1
python-dotenv==1.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
